Add Option to Get CRL when Calling GET CA REST API
This change adds a query parameter to the REST GET CA API, withCRL,
which MutentCA will use to return both the CA certificate and CRL, or
just the CA certificate.

This change is needed for the domain join with MutentCA work as we need
to synchronize both the trusted roots and CRLs for a particular CA.

Also fixes a bug in _LwCARestReadURIParams where it did not account for
query parameters if there is a variable in the URI path.

Jira ID: CNA-3599

Change-Id: I17cc304cdf348a6c8a8db0a33c6c8b042dad08dc