Change MutentCA Get CA API Response Object
MutentCA REST API to get CA details (certificate and CRL) has new
response object format:

```
{
    "caDetails": [
        { "cert": "...", "crls": "..." },
        { "crl": "...", "crls": "..." }
    ]
}
```

The key `crl` in each object will only be returned if the detail query
param is set to true.

This change is made to ensure a 1:1 mapping between CA certificate and
CRL.  If a CA has multiple certificates--e.g. during maintainence
mode--then each certificate will have a CRL, so there is an array of
pairings.

Change-Id: If5102ddb3ddb5f5ce36c0d623a05b85d8125447f