Support returning a secondary issuer in token claim
STS is configured with one issuer, configured by lwregshell and directory.
This change adds functionality to return another issuer value matching a
preset configuration in lwregshell if the request URI matches. If STS is
not configured to use secondary issuer, or if the request does not match
the secondary OIDC issuer, then the original behavior proceeds.

Testing:
- Ran integration and unit tests
- Manual testing: requested tokens using old endpoint and new endpoint
(both pointing to same STS) and configured the secondary issuer to be
new endpoint. Confirmed the issuer claim in token matched the request
URI

Change-Id: I243eca7cc9672618c63566405eb572b1c616e0c1