Refactor MutentCA AuthZ API to Pass Any X509 Struct
Previously, the LwCAAuthZCheckAccess function only accepted X509_REQ for
all APIs to check.  For revokcation APIs, X509_REQ is not the correct
type, instead, it should be X509_CRL, and in other cases, the data type
should be X509.

This change updates the AuthZ API to instead pass a union of each
desired X509 struct.  A union is used to keep the LwCAAuthZCheckAccess
API the same to avoid code duplication.  Also, LwCAAuthZCheckAccess
calls have been added to certificate and CA revokcation APIs.  For
plugins that do not implement AuthZ checks for those APIs, then it is
implied that AuthZ is not needed and it will be a no-op returning
authorized.

Jira ID: CNA-3851

Change-Id: I5277edf083436f88fda5e41a9bf961bd46669d58